pythonEnv/pyproject.toml
[tool.poetry]
name = "pythonenv"
version = "0.1.0"
description = ""
authors = ["Drixitel <mpmunoz1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"
numpy = "^1.24.2"
jupyter = "^1.0.0"
notebook = "^6.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
